+++ executors/kleverchain/scCallsExecutor_test.go
@@ -1,4 +1,4 @@
-package multiversx
+package kleverchain
 
 import (
 	"bytes"
